overlay: move overlay to built-in module
Problem: overlay runs in the broker thread which inhibits performance.

Move the overlay code to src/modules/overlay.
Add a module main() that is merely a stub.
Add module to broker built-ins list.

This breaks the overlay (and git bisect).
The commits for restoring functionality are kept separate for review.
