+++ src/modules/overlay/main.c
@@ -1,0 +1,32 @@
+/************************************************************\
+ * Copyright 2025 Lawrence Livermore National Security, LLC
+ * (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+ *
+ * This file is part of the Flux resource manager framework.
+ * For details, see https://github.com/flux-framework.
+ *
+ * SPDX-License-Identifier: LGPL-3.0
+\************************************************************/
+
+#if HAVE_CONFIG_H
+#include "config.h"
+#endif
+#include <flux/core.h>
+
+#include "src/broker/module.h"
+
+static int mod_main (flux_t *h, int argc, char *argv[])
+{
+    if (flux_reactor_run (flux_get_reactor (h), 0) < 0) {
+        flux_log_error (h, "flux_reactor_run");
+        return -1;
+    }
+    return 0;
+}
+
+struct module_builtin builtin_overlay = {
+    .name = "overlay",
+    .main = mod_main,
+};
+
+// vi:ts=4 sw=4 expandtab
